fix(command): refactor IMessage in order to fix undefined constructor bug in Command
