+++ libs/packages/ddd/src/application/__mocks__/commands.stub.ts
@@ -1,0 +1,4 @@
+import { Command, Props } from '../../domain'
+import { AggregateStub } from '../../domain/__mocks__'
+
+export class CreateAggregateStub extends Command<Props<AggregateStub>> {}
