+++ libs/packages/ddd/src/application/__mocks__/index.ts
@@ -1,1 +1,2 @@
+export * from './commands.stub'
 export * from './event-subscriber.mock'
