--- libs/packages/ddd/src/application/index.ts
@@ -1,2 +1,0 @@
-export * from './event/event-subscriber.interface'
-export * from './command/command.abstract'
