+++ libs/packages/ddd/src/domain/command/command-bus.interface.ts
@@ -1,0 +1,5 @@
+import { ICommand } from './command.abstract'
+
+export interface ICommandBus {
+  send(command: ICommand): Promise<void>
+}
