+++ libs/packages/ddd/src/domain/command/command-handler.interface.ts
@@ -1,0 +1,6 @@
+import { CommandClass, ICommand } from './command.abstract'
+
+export interface ICommandHandler<C extends ICommand = ICommand> {
+  subscription: CommandClass
+  handle(command: C): Promise<void>
+}
