+++ libs/packages/ddd/src/domain/command/command.abstract.ts
@@ -1,26 +1,25 @@
-import {
-  Constructor,
-  IMessage,
-  IMessageFactory,
-  IPayload,
-  Payload,
-} from '../../domain'
+import { Constructor, IMessage, IPayload, Payload } from '../../domain'
 
-export abstract class ICommand<
-  P extends IPayload = IPayload
-> extends IMessage<P> {
+export abstract class Command<P extends IPayload = IPayload>
+  implements ICommand<P>
+{
+  constructor(public readonly payload: P) {}
   /**
-   * Creates a new command from payload
+   * Creates a new DomainEvent from payload and aggregateId
    * @factory
    */
-  static override with<C extends IMessage = ICommand>(payload: Payload<C>): C {
-    return super.with<C>(payload)
+  static with<C extends ICommand = ICommand>(payload: Payload<C>): C {
+    return Reflect.construct(this, [payload])
   }
 }
 
 // eslint-disable-next-line @typescript-eslint/no-empty-interface
-export interface ICommandFactory<P extends IPayload = IPayload>
-  extends IMessageFactory<P, ICommand<P>> {}
+export interface ICommand<P extends IPayload = IPayload> extends IMessage<P> {}
+
+export interface ICommandFactory<P extends IPayload = IPayload> {
+  // eslint-disable-next-line @typescript-eslint/no-explicit-any
+  with: (payload: P) => any
+}
 
 export type CommandClass<P extends IPayload = IPayload> = Constructor &
   ICommandFactory<P>
