+++ libs/packages/ddd/src/domain/event/event-bus.interface.ts
@@ -1,5 +1,5 @@
-import { IEventSubscriber } from '../../application/event/event-subscriber.interface'
 import { IDomainEvent } from './domain-event.interface'
+import { IEventSubscriber } from './event-subscriber.interface'
 
 export interface IEventBus {
   publish(events: IDomainEvent[]): Promise<void>
