+++ libs/packages/ddd/src/domain/event/event-subscriber.interface.ts
@@ -1,4 +1,4 @@
-import { DomainEventClass, IDomainEvent } from '../../domain'
+import { DomainEventClass, IDomainEvent } from '..'
 
 export interface IEventSubscriber<E extends IDomainEvent = IDomainEvent> {
   subscriptions: DomainEventClass[]
