+++ libs/packages/ddd/src/domain/index.ts
@@ -5,6 +5,10 @@ export * from './core/errors'
 export * from './core/result'
 export * from './core/result.interface'
 
+export * from './command/command-bus.interface'
+export * from './command/command.abstract'
+export * from './command/command-handler.interface'
+
 export * from './entity/entity'
 export * from './entity/entity.abstract'
 export * from './entity/utils/unique.decorator'
@@ -13,6 +17,7 @@ export * from './event/domain-event.abstract'
 export * from './event/domain-event-id.vo'
 export * from './event/domain-event.interface'
 export * from './event/event-bus.interface'
+export * from './event/event-subscriber.interface'
 export * from './event/utils/domain-event.decorator'
 
 export * from './repo/entity-repo.abstract'
