+++ libs/packages/ddd/src/infra/__mocks__/event-bus.mock.ts
@@ -1,6 +1,4 @@
-import { IDomainEvent } from '../../domain/event/domain-event.interface'
-import { IEventBus } from '../../domain/event/event-bus.interface'
-import { IEventSubscriber } from '../../application'
+import { IDomainEvent, IEventBus, IEventSubscriber } from '../../domain'
 
 export class EventBusMock implements IEventBus {
   private _registerMock: jest.Mock = jest.fn()
