+++ libs/packages/ddd/src/infra/messaging/event-bus/in-memory-async.eventbus.ts
@@ -1,6 +1,5 @@
 import { EventEmitter } from 'events'
-import { IDomainEvent, IEventBus } from '../../../domain'
-import { IEventSubscriber } from '../../../application'
+import { IDomainEvent, IEventBus, IEventSubscriber } from '../../../domain'
 
 export class InMemoryAsyncEventBus extends EventEmitter implements IEventBus {
   register(subscribers: IEventSubscriber[]): void {
