+++ libs/packages/ddd/src/infra/messaging/event-bus/spec/in-memory-async.eventbus.spec.ts
@@ -1,14 +1,13 @@
-import {
-  EventSubscriberMock,
-  OtherEventSubscriberMock,
-} from '../../../../application/__mocks__'
+import { IUniqueID, IEventSubscriber } from '../../../../domain'
 import {
   AggregateCreatedStub,
   AggregateToggledStub,
   PropsUpdatedStub,
 } from '../../../../domain/__mocks__'
-import { IUniqueID } from '../../../../domain'
-import { IEventSubscriber } from '../../../../application'
+import {
+  EventSubscriberMock,
+  OtherEventSubscriberMock,
+} from '../../../../application/__mocks__'
 import { InMemoryAsyncEventBus } from '../in-memory-async.eventbus'
 
 describe(InMemoryAsyncEventBus, () => {
